MongoDbContextModel to Models folder MongoDbSetValidationModel to Models folder
