+++ MongoFramework.Tests/Models/MongoDbContextModel.cs
@@ -1,0 +1,6 @@
+﻿namespace MongoFramework.Tests.Models {
+	public class MongoDbContextModel
+	{
+		public string Id { get; set; }
+	}
+}
