+++ MongoFramework.Tests/MongoDbContextTests.cs
@@ -1,22 +1,9 @@
 ﻿using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using MongoFramework.Tests.Models;
 
 namespace MongoFramework.Tests
 {
-	public class MongoDbContextTestContext : MongoDbContext
-	{
-		public MongoDbContextTestContext(string connectionString, string databaseName) : base(connectionString, databaseName)
-		{
-		}
-
-		public MongoDbSet<MongoDbContextModel> ContextDbSet { get; set; }
-	}
-
-	public class MongoDbContextModel
-	{
-		public string Id { get; set; }
-	}
-
 	[TestClass]
 	public class MongoDbContextTests
 	{
